+++ GHEtool/Methods/__init__.py
@@ -1,3 +1,4 @@
 from .optimise_load_profile import optimise_load_profile_power, optimise_load_profile_energy, \
     optimise_load_profile_balance
 from .optimise_borefield_configuration import *
+from .optimise_load_profile_lp import optimise_load_profile_lp